destroy the Gallery database instance only if the user exits the app

+++ app/src/main/kotlin/com/simplemobiletools/gallery/activities/MainActivity.kt
@@ -190,7 +190,9 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
         config.temporarilyShowHidden = false
         mTempShowHiddenHandler.removeCallbacksAndMessages(null)
         removeTempFolder()
-        GalleryDataBase.destroyInstance()
+        if (!isChangingConfigurations) {
+            GalleryDataBase.destroyInstance()
+        }
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
